refactor(*): reorganize folder structure
