+++ odrc/algorithm/width-check.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <odrc/interface/gdsii/gdsii.hpp>
+#include <odrc/gdsii/gdsii.hpp>
 
 namespace odrc {
     void width_check(const odrc::gdsii::library &lib, int layer, int threshold);
