--- odrc/infrastructure/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_subdirectory(exception)
-add_subdirectory(logger)
-add_subdirectory(sweepline)
