--- odrc/infrastructure/exception/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_library(libexception INTERFACE)
-target_include_directories(libexception INTERFACE ${CMAKE_SOURCE_DIR})
