--- odrc/infrastructure/logger/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_library(liblogger INTERFACE)
-target_include_directories(liblogger INTERFACE ${CMAKE_SOURCE_DIR} ${CMAKE_SOURCE_DIR}/thirdparty)
