--- odrc/infrastructure/task-queue.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <queue>
-
-namespace odrc::core {
-    class task_queue {
-
-    };
-}
