+++ tests/core/sweepline.test.cpp
@@ -1,10 +1,10 @@
-#include <odrc/infrastructure/sweepline/sweepline.hpp>
+#include <odrc/core/sweepline.hpp>
 
 #include <vector>
 
 #include <doctest/doctest.h>
 
-#include <odrc/infrastructure/execution.hpp>
+#include <odrc/core/execution.hpp>
 
 TEST_SUITE("[OpenDRC] odrc::sweepline tests") {
   TEST_CASE("test trivial sequenced copy") {
